+++ relay-event-normalization/src/event.rs
@@ -1514,6 +1514,7 @@ fn normalize_app_start_measurements(measurements: &mut Measurements) {
 #[cfg(test)]
 mod tests {
 
+    use relay_pattern::Pattern;
     use std::collections::BTreeMap;
     use std::collections::HashMap;
 
@@ -2285,7 +2286,7 @@ mod tests {
                     version: 2,
                     models: HashMap::from([
                         (
-                            "claude-2.1".to_owned(),
+                            Pattern::new("claude-2.1").unwrap(),
                             ModelCostV2 {
                                 input_per_token: 0.01,
                                 output_per_token: 0.02,
@@ -2294,7 +2295,7 @@ mod tests {
                             },
                         ),
                         (
-                            "gpt4-21-04".to_owned(),
+                            Pattern::new("gpt4-21-04").unwrap(),
                             ModelCostV2 {
                                 input_per_token: 0.02,
                                 output_per_token: 0.03,
@@ -2323,6 +2324,14 @@ mod tests {
                 .first()
                 .and_then(|span| span.value())
                 .and_then(|span| span.data.value())
+                .and_then(|data| data.gen_ai_cost_total_tokens.value()),
+            Some(&Value::F64(50.0))
+        );
+        assert_eq!(
+            spans
+                .first()
+                .and_then(|span| span.value())
+                .and_then(|span| span.data.value())
                 .and_then(|data| data.gen_ai_cost_input_tokens.value()),
             Some(&Value::F64(10.0))
         );
@@ -2340,6 +2349,14 @@ mod tests {
                 .and_then(|span| span.value())
                 .and_then(|span| span.data.value())
                 .and_then(|data| data.gen_ai_usage_total_cost.value()),
+            Some(&Value::F64(80.0))
+        );
+        assert_eq!(
+            spans
+                .get(1)
+                .and_then(|span| span.value())
+                .and_then(|span| span.data.value())
+                .and_then(|data| data.gen_ai_cost_total_tokens.value()),
             Some(&Value::F64(80.0))
         );
         assert_eq!(
@@ -2422,7 +2439,7 @@ mod tests {
                     version: 2,
                     models: HashMap::from([
                         (
-                            "claude-2.1".to_owned(),
+                            Pattern::new("claude-2.1").unwrap(),
                             ModelCostV2 {
                                 input_per_token: 0.01,
                                 output_per_token: 0.02,
@@ -2431,7 +2448,7 @@ mod tests {
                             },
                         ),
                         (
-                            "gpt4-21-04".to_owned(),
+                            Pattern::new("gpt4-21-04").unwrap(),
                             ModelCostV2 {
                                 input_per_token: 0.09,
                                 output_per_token: 0.05,
@@ -2456,6 +2473,10 @@ mod tests {
             Some(&Value::F64(75.0))
         );
         assert_eq!(
+            first_span_data.and_then(|data| data.gen_ai_cost_total_tokens.value()),
+            Some(&Value::F64(75.0))
+        );
+        assert_eq!(
             first_span_data.and_then(|data| data.gen_ai_cost_input_tokens.value()),
             Some(&Value::F64(25.0))
         );
@@ -2474,6 +2495,10 @@ mod tests {
             .and_then(|span| span.data.value());
         assert_eq!(
             second_span_data.and_then(|data| data.gen_ai_usage_total_cost.value()),
+            Some(&Value::F64(190.0))
+        );
+        assert_eq!(
+            second_span_data.and_then(|data| data.gen_ai_cost_total_tokens.value()),
             Some(&Value::F64(190.0))
         );
         assert_eq!(
@@ -2500,6 +2525,10 @@ mod tests {
             .and_then(|span| span.data.value());
         assert_eq!(
             third_span_data.and_then(|data| data.gen_ai_usage_total_cost.value()),
+            Some(&Value::F64(190.0))
+        );
+        assert_eq!(
+            third_span_data.and_then(|data| data.gen_ai_cost_total_tokens.value()),
             Some(&Value::F64(190.0))
         );
         assert_eq!(
@@ -2541,7 +2570,7 @@ mod tests {
                 ai_model_costs: Some(&ModelCosts {
                     version: 2,
                     models: HashMap::from([(
-                        "claude-2.1".to_owned(),
+                        Pattern::new("claude-2.1").unwrap(),
                         ModelCostV2 {
                             input_per_token: 0.01,
                             output_per_token: 0.02,
@@ -2564,6 +2593,14 @@ mod tests {
                 .and_then(|span| span.value())
                 .and_then(|span| span.data.value())
                 .and_then(|data| data.gen_ai_usage_total_cost.value()),
+            None
+        );
+        assert_eq!(
+            spans
+                .first()
+                .and_then(|span| span.value())
+                .and_then(|span| span.data.value())
+                .and_then(|data| data.gen_ai_cost_total_tokens.value()),
             None
         );
         // total_tokens shouldn't be set if no tokens are present on span data
@@ -2625,7 +2662,7 @@ mod tests {
                     version: 2,
                     models: HashMap::from([
                         (
-                            "claude-2.1".to_owned(),
+                            Pattern::new("claude-2.1").unwrap(),
                             ModelCostV2 {
                                 input_per_token: 0.01,
                                 output_per_token: 0.02,
@@ -2634,7 +2671,7 @@ mod tests {
                             },
                         ),
                         (
-                            "gpt4-21-04".to_owned(),
+                            Pattern::new("gpt4-21-04").unwrap(),
                             ModelCostV2 {
                                 input_per_token: 0.09,
                                 output_per_token: 0.05,
@@ -2663,6 +2700,14 @@ mod tests {
                 .first()
                 .and_then(|span| span.value())
                 .and_then(|span| span.data.value())
+                .and_then(|data| data.gen_ai_cost_total_tokens.value()),
+            Some(&Value::F64(65.0))
+        );
+        assert_eq!(
+            spans
+                .first()
+                .and_then(|span| span.value())
+                .and_then(|span| span.data.value())
                 .and_then(|data| data.gen_ai_cost_input_tokens.value()),
             Some(&Value::F64(25.0))
         );
@@ -2680,6 +2725,14 @@ mod tests {
                 .and_then(|span| span.value())
                 .and_then(|span| span.data.value())
                 .and_then(|data| data.gen_ai_usage_total_cost.value()),
+            Some(&Value::F64(190.0))
+        );
+        assert_eq!(
+            spans
+                .get(1)
+                .and_then(|span| span.value())
+                .and_then(|span| span.data.value())
+                .and_then(|data| data.gen_ai_cost_total_tokens.value()),
             Some(&Value::F64(190.0))
         );
         assert_eq!(